move volume/packmol; update flake setting
