+++ mdgo/util/__init__.py
@@ -1,0 +1,14 @@
+# coding: utf-8
+# Copyright (c) Tingzheng Hou.
+# Distributed under the terms of the MIT License.
+
+"""
+The util package implements various utilities that are commonly used by various
+packages.
+"""
+
+__author__ = "Tingzheng Hou"
+__version__ = "0.3.0"
+__maintainer__ = "Tingzheng Hou"
+__email__ = "tingzheng_hou@berkeley.edu"
+__date__ = "Jul 19, 2021"
