+++ setup.cfg
@@ -23,6 +23,9 @@ exclude = .git,__pycache__,docs_rst/conf.py,tests
 # max-complexity = 10
 extend-ignore = E741,W291,W293,E501,E231,E203
 max-line-length = 120
+per-file-ignores =
+  # F401: imported but unused
+  __init__.py: F401
 
 [pydocstyle]
 ignore = D103,D105,D2,D4
